+++ .env.production
@@ -1,0 +1,1 @@
+VITE_API_BASE_URL=http://54.180.165.220/api
