+++ src/components/common/modal/Modal.module.scss
@@ -117,3 +117,7 @@
 .active {
   background-color: lightgray;
 }
+
+.errorMsg {
+  color: red;
+}
