+++ src/pages/editprofile/styles/editprofile.module.scss
@@ -18,112 +18,7 @@
     overflow: scroll;
 
     color: $color-purple-500;
-
-    &__infoBox {
-      width: 100%;
-      display: flex;
-      flex-direction: column;
-      justify-content: space-between;
-      gap: 3rem;
-
-      padding: 1rem;
-      border-radius: 15px;
-      box-shadow: 0px 4px 8px $color-gray-200;
-
-      &__title {
-        font-size: 1.5rem;
-      }
-
-      &__body {
-        display: flex;
-        flex-direction: column;
-        gap: 20px;
-
-        &__editId {
-          display: flex;
-          flex-direction: column;
-          gap: 10px;
-        }
-
-        &__editPw {
-          display: flex;
-          flex-direction: column;
-          gap: 10px;
-        }
-
-        &__editName {
-          display: flex;
-          flex-direction: column;
-          gap: 10px;
-        }
-      }
-    }
-
-    &__preferBox {
-      width: 100%;
-      display: flex;
-      flex-direction: column;
-      justify-content: space-between;
-      gap: 3rem;
-
-      padding: 1rem;
-      border-radius: 15px;
-      box-shadow: 0px 4px 8px $color-gray-200;
-
-      &__title {
-        font-size: 1.5rem;
-      }
-      &__body {
-        display: flex;
-        flex-direction: column;
-        gap: 20px;
-
-        &__editParty {
-          display: flex;
-          flex-direction: column;
-          gap: 10px;
-          &__text {
-            width: 100%;
-            height: 3.5rem;
-            padding-left: 10px;
-            padding-right: 10px;
-
-            border-style: solid;
-            border-width: 2px;
-            border-color: $color-purple-500;
-            border-radius: 10px;
-          }
-        }
-        &__editPolitician {
-          display: flex;
-          flex-direction: column;
-          gap: 10px;
-
-          &__text {
-            width: 100%;
-            height: 3.5rem;
-            padding-left: 10px;
-            padding-right: 10px;
-
-            border-style: solid;
-            border-width: 2px;
-            border-color: $color-purple-500;
-            border-radius: 10px;
-          }
-        }
-      }
-    }
   }
-}
-
-.errorMsg {
-  color: red;
-  font-size: 0.7rem;
-}
-
-.noErrorMsg {
-  visibility: hidden;
-  font-size: 0.7rem;
 }
 
 .overlay {
