+++ src/pages/main/personalview/PersonalView.tsx
@@ -1,110 +1,15 @@
-//import React from "react";
 import TopNav from "@/components/common/TopNav/TopNav";
 import styles from "./styles/personalview.module.scss";
 import BotNav from "@/components/common/botnav/BotNav";
-import { useEffect, useState } from "react";
 import { MdKeyboardArrowRight } from "react-icons/md";
-import { useNavigate } from "react-router-dom";
-import { useRecoilState } from "recoil";
-import { curMainNum } from "@/recoil/atoms/curMainNum";
+import ProfileBox from "./profilebox/ProfileBox";
 
 function PersonalView() {
-  const navigate = useNavigate();
-  const [intPol, setIntPol] = useState<string>("");
-  const [intPar, setIntPar] = useState<string>("");
-  const [curView, setCurView] = useRecoilState(curMainNum);
-
-  const logout = () => {
-    sessionStorage.removeItem("access-token");
-    sessionStorage.removeItem("refresh-token");
-    sessionStorage.removeItem("nickname");
-    sessionStorage.removeItem("politicianOfInterest");
-    sessionStorage.removeItem("partyOfInterest");
-    setCurView(1);
-    navigate("/");
-  };
-
-  useEffect(() => {
-    setIntPol(sessionStorage.getItem("politicianOfInterest"));
-    setIntPar(sessionStorage.getItem("partyOfInterest"));
-  }, []);
-
   return (
     <div className={styles.page}>
       <TopNav></TopNav>
       <div className={styles.page__contents}>
-        <div className={styles.page__contents__profileSection}>
-          <div className={styles.page__contents__profileSection__name}>
-            {sessionStorage.getItem("nickname")}님
-          </div>
-          <div className={styles.page__contents__profileSection__dataBox}>
-            <div className={styles.page__contents__profileSection__dataBox__prefer}>
-              <div className={styles.page__contents__profileSection__dataBox__prefer__party}>
-                <div
-                  className={styles.page__contents__profileSection__dataBox__prefer__party__label}
-                >
-                  선호 정당
-                </div>
-                <div
-                  className={styles.page__contents__profileSection__dataBox__prefer__party__data}
-                >
-                  {intPar === "" ? "미등록" : intPar}
-                </div>
-              </div>
-              <div className={styles.page__contents__profileSection__dataBox__prefer__politician}>
-                <div
-                  className={
-                    styles.page__contents__profileSection__dataBox__prefer__politician__label
-                  }
-                >
-                  선호 정치인
-                </div>
-                <div
-                  className={
-                    styles.page__contents__profileSection__dataBox__prefer__politician__data
-                  }
-                >
-                  {intPol === "" ? "미등록" : intPol}
-                </div>
-              </div>
-            </div>
-            <div className={styles.page__contents__profileSection__dataBox__info}>
-              <div
-                className={styles.page__contents__profileSection__dataBox__info__edit}
-                onClick={() => navigate("/edit")}
-              >
-                <div>내 정보 수정</div>
-                <div>
-                  <MdKeyboardArrowRight />
-                </div>
-              </div>
-              <div
-                className={styles.page__contents__profileSection__dataBox__info__logout}
-                onClick={logout}
-              >
-                로그아웃
-              </div>
-            </div>
-          </div>
-          <div className={styles.page__contents__profileSection__status}>
-            <div className={styles.page__contents__profileSection__status__post}>
-              <div className={styles.page__contents__profileSection__status__post__totalCount}>
-                내가쓴글 8개
-              </div>
-              <div className={styles.viewAll}>
-                <div>전체보기</div>
-                <MdKeyboardArrowRight />
-              </div>
-            </div>
-            {/* <div className={styles.page__contents__profileSection__status__donate}>
-              <div className={styles.page__contents__profileSection__status__donate__totalCount}>후원중인 정치인 8명</div>
-              <div className={styles.viewAll}>
-                <div>전체보기</div>
-                <MdKeyboardArrowRight />
-              </div>
-            </div> */}
-          </div>
-        </div>
+        <ProfileBox />
         <div className={styles.page__contents__recordSection}>
           <div className={styles.page__contents__recordSection__scrapContainer}>
             <div className={styles.page__contents__recordSection__scrapContainer__head}>
