+++ src/pages/politician/PoliticianInfo.tsx
@@ -18,9 +18,7 @@ function PoliticianInfo() {
   const [curPos, setCurPos] = useState(0);
   // const [curPos, setCurPos] = useState(0);
 
-  // const ENDPOINT = `https://d282ffdd-b1e5-4e5a-bebc-2a161c592cb5.mock.pstmn.io/politician/${params.id}`;
-  // const ENDPOINT = `http://localhost:9001/politicians/${params.id}/detail`;
-  const ENDPOINT = `http://54.180.165.220/api/politicians/${params.id}/detail`;
+  const PATH = `/politicians/${params.id}/detail`;
 
   /*지켜보기 기능*/
   const [observe, setObserve] = useState<boolean>(false);
@@ -156,7 +154,7 @@ function PoliticianInfo() {
   useEffect(() => {
     // console.log(params);
     const getPoliticanData = async () => {
-      const result = await apiCall(ENDPOINT, "GET");
+      const result = await apiCall(PATH, "GET");
 
       if (result.code === 20000) {
         console.log(result.message);
