+++ src/recoil/selectors/eventSelector.ts
@@ -1,16 +1,13 @@
 import { selector } from "recoil";
 import { yearState } from "../atoms/year";
 
-// const ENDPOINT = "https://d282ffdd-b1e5-4e5a-bebc-2a161c592cb5.mock.pstmn.io/calendar/schedules?year=${yearValue}"
-// const ENDPOINT = `http://localhost:9001/calendar/schedules?userId=${userId}&year=${yearValue}`
 const userId = sessionStorage.getItem("id");
 
 export const eventData = selector({
   key: "eventData",
   get: async ({ get }) => {
     const yearValue = get(yearState);
-    // const ENDPOINT = `http://localhost:9001/calendar/schedules?userId=${userId}&year=${yearValue}`;
-    const ENDPOINT = `http://54.180.165.220/api/calendar/schedules?userId=${userId}&year=${yearValue}`;
+    const PATH = `/calendar/schedules?userId=${userId}&year=${yearValue}`;
     const access_token = sessionStorage.getItem("access-token");
 
     const headers = {
@@ -19,7 +16,7 @@ export const eventData = selector({
     };
 
     try {
-      const response = await fetch(ENDPOINT, {
+      const response = await fetch(PATH, {
         method: "GET",
         headers,
       });
